% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accidents}
\alias{accidents}
\title{accidents}
\format{
A data frame with 600 observations and 11 variables:
\describe{
\item{RushHour}{}
\item{WRK_ZONE}{}
\item{WKDY}{}
\item{INT_HWY}{}
\item{LGTCON_day}{}
\item{LEVEL}{}
\item{SPD_LIM}{}
\item{SUR_COND_dry}{}
\item{TRAF_two_way}{}
\item{WEATHER_adverse}{}
\item{MAX_SEV}{}
}
}
\usage{
accidents
}
\description{
Data
}
\examples{
accidents
}
\keyword{datasets}
