% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Airfares}
\alias{Airfares}
\title{Airfares}
\format{
A data frame with 638 observations and 18 variables:
\describe{
\item{S_CODE}{}
\item{S_CITY}{}
\item{E_CODE}{}
\item{E_CITY}{}
\item{COUPON}{}
\item{NEW}{}
\item{VACATION}{}
\item{SW}{}
\item{HI}{}
\item{S_INCOME}{}
\item{E_INCOME}{}
\item{S_POP}{}
\item{E_POP}{}
\item{SLOT}{}
\item{GATE}{}
\item{DISTANCE}{}
\item{PAX}{}
\item{FARE}{}
}
}
\usage{
Airfares
}
\description{
Data
}
\examples{
head(Airfares)
}
\keyword{datasets}
