% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Amtrak}
\alias{Amtrak}
\title{Amtrak}
\format{
A data frame with 159 observations and 2 variables:
\describe{
\item{Month}{}
\item{Ridership}{Amtrak Ridership Number of Passengers (in thousands) }
}
}
\usage{
Amtrak
}
\description{
Data
}
\examples{
Amtrak
}
\keyword{datasets}
