% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ApplianceShipments}
\alias{ApplianceShipments}
\title{Applicance Shipments}
\format{
A data frame with 20 observations and 2 variables:
\describe{
\item{Quarter}{}
\item{Shipments}{}
}
}
\usage{
ApplianceShipments
}
\description{
Data
}
\examples{
ApplianceShipments
}
\keyword{datasets}
