% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AustralianWines}
\alias{AustralianWines}
\title{AustralianWines}
\format{
A data frame with 180 observations and 7 variables:
\describe{
\item{month}{}
\item{fortified}{}
\item{red}{}
\item{rose}{}
\item{sparkling}{}
\item{sweet.white}{}
\item{dry.white}{}
}
}
\usage{
AustralianWines
}
\description{
Data
}
\examples{
AustralianWines
}
\keyword{datasets}
