% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{AutosElectronics}
\alias{AutosElectronics}
\title{AutosElectronics}
\format{
An object of class \code{character} of length 1.
}
\source{
{The posts are taken from Internet groups devoted to autos and electronics}
}
\usage{
AutosElectronics
}
\value{
the filename of the dataset that is installed with the mlba package
}
\description{
{Filename of data set for a classification task—to classify Internet discussion
posts as either auto-related or electronics-related.

The data set contains two sets of messages; one with 1000 auto-related posts and one with 1000 electronics-related posts.}
}
\examples{
\dontrun{corpus <- Corpus(ZipSource(AutosElectronics, recursive=True))}
}
\keyword{datasets}
