% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bankruptcy}
\alias{Bankruptcy}
\title{Bankruptcy}
\format{
A data frame with 132 observations and 27 variables:
\describe{
\item{NO}{}
\item{D}{}
\item{YR}{}
\item{R1}{}
\item{R2}{}
\item{R3}{}
\item{R4}{}
\item{R5}{}
\item{R6}{}
\item{R7}{}
\item{R8}{}
\item{R9}{}
\item{R10}{}
\item{R11}{}
\item{R12}{}
\item{R13}{}
\item{R14}{}
\item{R15}{}
\item{R16}{}
\item{R17}{}
\item{R18}{}
\item{R19}{}
\item{R20}{}
\item{R21}{}
\item{R22}{}
\item{R23}{}
\item{R24}{}
}
}
\usage{
Bankruptcy
}
\description{
Data
}
\examples{
head(Bankruptcy)
}
\keyword{datasets}
