% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{banks}
\alias{banks}
\title{Banks}
\format{
A data frame with 20 observations and 5 variables:
\describe{
\item{Obs}{}
\item{Financial.Condition}{}
\item{TotCap.Assets}{}
\item{TotExp.Assets}{}
\item{TotLns.Lses.Assets}{}
}
}
\usage{
banks
}
\description{
Data
}
\examples{
banks
}
\keyword{datasets}
