% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BareggTunnel}
\alias{BareggTunnel}
\title{BareggTunnel}
\format{
A data frame with 747 observations and 2 variables:
\describe{
\item{Day}{}
\item{Number.of.vehicles}{}
}
}
\usage{
BareggTunnel
}
\description{
Data
}
\examples{
BareggTunnel
}
\keyword{datasets}
