% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bicup2006}
\alias{bicup2006}
\title{bicup2006}
\format{
A data frame with 1512 observations and 3 variables:
\describe{
\item{DATE}{}
\item{TIME}{}
\item{DEMAND}{}
}
}
\source{
{Ken Black (used by permission)}
}
\usage{
bicup2006
}
\description{
Data
}
\examples{
bicup2006
}
\keyword{datasets}
