% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BostonHousing}
\alias{BostonHousing}
\title{Boston Housing}
\format{
A data frame with 506 observations and 14 variables:
\describe{
\item{CRIM}{}
\item{ZN}{}
\item{INDUS}{}
\item{CHAS}{}
\item{NOX}{}
\item{RM}{}
\item{AGE}{}
\item{DIS}{}
\item{RAD}{}
\item{TAX}{}
\item{PTRATIO}{}
\item{LSTAT}{}
\item{MEDV}{}
\item{CAT.MEDV}{}
}
}
\usage{
BostonHousing
}
\description{
Data
}
\examples{
BostonHousing
}
\keyword{datasets}
