% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cereals}
\alias{Cereals}
\title{Cereals}
\format{
A data frame with 77 observations and 16 variables:
\describe{
\item{name}{}
\item{mfr}{}
\item{type}{}
\item{calories}{}
\item{protein}{}
\item{fat}{}
\item{sodium}{}
\item{fiber}{}
\item{carbo}{}
\item{sugars}{}
\item{potass}{}
\item{vitamins}{}
\item{shelf}{}
\item{weight}{}
\item{cups}{}
\item{rating}{}
}
}
\usage{
Cereals
}
\description{
Data
}
\examples{
Cereals
}
\keyword{datasets}
