% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CharlesBookClub}
\alias{CharlesBookClub}
\title{Charles Book Club}
\format{
A data frame with 4000 observations and 24 variables:
\describe{
\item{Seq.}{}
\item{ID.}{}
\item{Gender}{}
\item{M}{}
\item{R}{}
\item{F}{}
\item{FirstPurch}{}
\item{ChildBks}{}
\item{YouthBks}{}
\item{CookBks}{}
\item{DoItYBks}{}
\item{RefBks}{}
\item{ArtBks}{}
\item{GeogBks}{}
\item{ItalCook}{}
\item{ItalAtlas}{}
\item{ItalArt}{}
\item{Florence}{}
\item{Related.Purchase}{}
\item{Mcode}{}
\item{Rcode}{}
\item{Fcode}{}
\item{Yes_Florence}{}
\item{No_Florence}{}
}
}
\usage{
CharlesBookClub
}
\description{
Data
}
\examples{
head(CharlesBookClub)
}
\keyword{datasets}
