% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COMPAS_clean}
\alias{COMPAS_clean}
\title{COMPAS clean}
\format{
A data frame with 5304 observations and 7 variables
\describe{
\item{id}{}
\item{age_cat}{categories: Less than 25, 25 - 45, Greater than 45}
\item{c_charge_degree}{}
\item{race}{}
\item{sex}{}
\item{priors_count}{}
\item{two_year_recid}{}
}
}
\source{
{https://github.com/propublica/compas-analysis/blob/master/compas-scores-two-years.csv}
}
\usage{
COMPAS_clean
}
\description{
Dataset derived from the ProPublica story "Machine Bias."
See https://github.com/propublica/compas-analysis for details
}
\details{
Data
}
\examples{
head(COMPAS_clean)
}
\keyword{datasets}
