% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CompasScoresTwoYears}
\alias{CompasScoresTwoYears}
\title{CompasScoresTwoYears}
\format{
A data frame with 7214 observations and 53 variables
}
\source{
{https://github.com/propublica/compas-analysis/blob/master/compas-scores-two-years.csv}
}
\usage{
CompasScoresTwoYears
}
\description{
Dataset from the ProPublica story "Machine Bias."
See https://github.com/propublica/compas-analysis for details
}
\details{
Data
}
\examples{
head(CompasScoresTwoYears)
}
\keyword{datasets}
