% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cosmetics}
\alias{Cosmetics}
\title{Cosmetics}
\format{
A data frame with 1000 observations and 15 variables:
\describe{
\item{Trans.}{}
\item{Bag}{}
\item{Blush}{}
\item{Nail.Polish}{}
\item{Brushes}{}
\item{Concealer}{}
\item{Eyebrow.Pencils}{}
\item{Bronzer}{}
\item{Lip.liner}{}
\item{Mascara}{}
\item{Eye.shadow}{}
\item{Foundation}{}
\item{Lip.Gloss}{}
\item{Lipstick}{}
\item{Eyeliner}{}
}
}
\usage{
Cosmetics
}
\description{
Data
}
\examples{
head(Cosmetics)
}
\keyword{datasets}
