% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{courserating}
\alias{courserating}
\title{Course Rating}
\format{
A data frame with 15 observations and 10 variables:
\describe{
\item{X}{}
\item{SQL}{}
\item{Spatial}{}
\item{PA1}{}
\item{DM.in.R}{}
\item{Python}{}
\item{Forecast}{}
\item{R.Prog}{}
\item{Hadoop}{}
\item{Regression}{}
}
}
\source{
{Copyright 2016 statistics.com}
}
\usage{
courserating
}
\description{
Student ratings of online statistics courses at Statistics.com
}
\examples{
head(courserating)
}
\keyword{datasets}
