% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Coursetopics}
\alias{Coursetopics}
\title{Coursetopics}
\format{
A data frame with 365 observations and 8 variables:
\describe{
\item{Intro}{}
\item{DataMining}{}
\item{Survey}{}
\item{Cat.Data}{}
\item{Regression}{}
\item{Forecast}{}
\item{DOE}{}
\item{SW}{}
}
}
\usage{
Coursetopics
}
\description{
Data
}
\examples{
head(Coursetopics)
}
\keyword{datasets}
