% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DepartmentStoreSales}
\alias{DepartmentStoreSales}
\title{DepartmentStoreSales}
\format{
A data frame with 24 observations and 2 variables:
\describe{
\item{Quarter}{}
\item{Sales}{}
}
}
\usage{
DepartmentStoreSales
}
\description{
Data
}
\examples{
DepartmentStoreSales
}
\keyword{datasets}
