% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Drug}
\alias{Drug}
\title{Drug}
\format{
A data frame with 60 observations and 7 variables:
\describe{
\item{Entity}{}
\item{Related.Entity}{}
\item{Relationship}{}
\item{Descrption}{}
\item{Related.Entity.Address.1}{}
\item{Related.Entity.Address.2}{}
\item{Related.Address.3}{}
}
}
\usage{
Drug
}
\description{
Data
}
\examples{
Drug
}
\keyword{datasets}
