% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EastWestAirlinesCluster}
\alias{EastWestAirlinesCluster}
\title{EastWestAirlinesCluster}
\format{
A data frame with 3999 observations and 12 variables:
\describe{
\item{ID.}{}
\item{Balance}{}
\item{Qual_miles}{}
\item{cc1_miles}{}
\item{cc2_miles}{}
\item{cc3_miles}{}
\item{Bonus_miles}{}
\item{Bonus_trans}{}
\item{Flight_miles_12mo}{}
\item{Flight_trans_12}{}
\item{Days_since_enroll}{}
\item{Award.}{}
}
}
\usage{
EastWestAirlinesCluster
}
\description{
Data
}
\examples{
head(EastWestAirlinesCluster)
}
\keyword{datasets}
