% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eBayAuctions}
\alias{eBayAuctions}
\title{ebay Auctions}
\format{
A data frame with 1972 observations and 8 variables:
\describe{
\item{Category}{}
\item{currency}{}
\item{sellerRating}{}
\item{Duration}{}
\item{endDay}{}
\item{ClosePrice}{}
\item{OpenPrice}{}
\item{Competitive.}{}
}
}
\usage{
eBayAuctions
}
\description{
Data
}
\examples{
eBayAuctions
}
\keyword{datasets}
