% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eBayNetwork}
\alias{eBayNetwork}
\title{ebay Network}
\format{
A data frame with 200 observations and 5 variables:
\describe{
\item{Seller}{}
\item{Bidder}{}
\item{Weight}{}
\item{Bidder.Volume}{}
\item{Seller.Volume}{}
}
}
\usage{
eBayNetwork
}
\description{
Data
}
\examples{
eBayNetwork
}
\keyword{datasets}
