% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eBayTreemap}
\alias{eBayTreemap}
\title{ebay Treemap}
\format{
A data frame with 10078 observations and 5 variables:
\describe{
\item{High.Bid}{}
\item{Seller.Feedback}{}
\item{Category}{}
\item{Sub.Category}{}
\item{Brand}{}
}
}
\usage{
eBayTreemap
}
\description{
Data
}
\examples{
head(eBayTreemap)
}
\keyword{datasets}
