% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EmailABtest}
\alias{EmailABtest}
\title{EmailABtest}
\format{
A data frame with 426 observations and 3 variables:
\describe{
\item{Customer}{}
\item{Message}{}
\item{Open.}{}
}
}
\usage{
EmailABtest
}
\description{
Data
}
\examples{
head(EmailABtest)
}
\keyword{datasets}
