% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FarmAds}
\alias{FarmAds}
\title{FarmAds}
\format{
A data frame with 4143 observations and 2 variables:
\describe{
\item{label}{}
\item{text}{}
}
}
\usage{
FarmAds
}
\description{
Data
}
\examples{
FarmAds
}
\keyword{datasets}
