% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FlightDelays}
\alias{FlightDelays}
\title{Flight delays}
\format{
A data frame with 2201 observations and 13 variables:
\describe{
\item{CRS_DEP_TIME}{}
\item{CARRIER}{}
\item{DEP_TIME}{}
\item{DEST}{}
\item{DISTANCE}{}
\item{FL_DATE}{}
\item{FL_NUM}{}
\item{ORIGIN}{}
\item{Weather}{}
\item{DAY_WEEK}{}
\item{DAY_OF_MONTH}{}
\item{TAIL_NUM}{}
\item{Flight.Status}{}
}
}
\usage{
FlightDelays
}
\description{
Data
}
\examples{
head(FlightDelays)
}
\keyword{datasets}
