% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FutureFundraising}
\alias{FutureFundraising}
\title{FutureFundraising}
\format{
A data frame with 2000 observations and 24 variables:
\describe{
\item{Row.Id}{}
\item{Row.Id.}{}
\item{zipconvert_2}{}
\item{zipconvert_3}{}
\item{zipconvert_4}{}
\item{zipconvert_5}{}
\item{homeowner.dummy}{}
\item{NUMCHLD}{}
\item{INCOME}{}
\item{gender.dummy}{}
\item{WEALTH}{}
\item{HV}{}
\item{Icmed}{}
\item{Icavg}{}
\item{IC15}{}
\item{NUMPROM}{}
\item{RAMNTALL}{}
\item{MAXRAMNT}{}
\item{LASTGIFT}{}
\item{totalmonths}{}
\item{TIMELAG}{}
\item{AVGGIFT}{}
\item{TARGET_B}{}
\item{TARGET_D}{}
}
}
\usage{
FutureFundraising
}
\description{
Data
}
\examples{
head(FutureFundraising)
}
\seealso{
{See \code{\link{Fundraising}} for a description of the dataset}
}
\keyword{datasets}
