% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{GDP}
\format{
A data frame with 264 observations and 5 variables:
\describe{
\item{Country.Name}{}
\item{Country.Code}{}
\item{Indicator.Name}{}
\item{Indicator.Code}{}
\item{GDP2015}{}
}
}
\usage{
gdp
}
\description{
Data
}
\examples{
head(gdp)
}
\keyword{datasets}
