% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GermanCredit}
\alias{GermanCredit}
\title{German credit}
\format{
A data frame with 1000 observations and 32 variables:
\describe{
\item{OBS.}{Observation No.}
\item{CHK_ACCT}{Checking account status. 0: < 0 DM; 1: 0 <...< 200 DM;
2: => 200 DM; 3:  no checking account}
\item{DURATION}{Duration of credit in months}
\item{HISTORY}{Credit history. 0: no credits taken;
1: all credits at this bank paid back duly;
2: existing credits paid back duly till now;
3: delay in paying off in the past;
4: critical account}
\item{NEW_CAR}{Purpose of credit. car (new) 0: No, 1: Yes}
\item{USED_CAR}{Purpose of credit. car (used) 0: No, 1: Yes}
\item{FURNITURE}{Purpose of credit. furniture/equipment 0: No, 1: Yes}
\item{RADIO.TV}{Purpose of credit. radio/television 0: No, 1: Yes}
\item{EDUCATION}{Purpose of credit. education 0: No, 1: Yes}
\item{RETRAINING}{Purpose of credit. retraining 0: No, 1: Yes}
\item{AMOUNT}{Credit amount}
\item{SAV_ACCT}{Average balance in savings account.
0 : <  100 DM;
1 : 100 <=...<  500 DM;
2 : 500 <=...< 1000 DM;
3 : =>1000 DM;
4 : unknown/ no savings account}
\item{EMPLOYMENT}{Present employment since.
0 : unemployed;
1 :  < 1 year;
2 : 1 <= ... < 4 years
3 : 4 <=... < 7 years
4 : >= 7 years}
\item{INSTALL_RATE}{Installment rate as \% of disposable income}
\item{MALE_DIV}{Applicant is male and divorced. 0: No, 1: Yes}
\item{MALE_SINGLE}{Applicant is male and single. 0: No, 1: Yes}
\item{MALE_MAR_or_WID}{Applicant is male and married or a widower. 0: No, 1: Yes}
\item{CO.APPLICANT}{Application has a co-applicant. 0: No, 1: Yes}
\item{GUARANTOR}{Applicant has a guarantor. 0: No, 1: Yes}
\item{PRESENT_RESIDENT}{Present resident since-years.
0: <= 1 year;
1: <…<= 2 years;
2: <…<= 3 years;
3: >4years}
\item{REAL_ESTATE}{Applicant owns real estate. 0: No, 1: Yes}
\item{PROP_UNKN_NONE}{Applicant owns no property (or unknown). 0: No, 1: Yes}
\item{AGE}{Age in years}
\item{OTHER_INSTALL}{Applicant has other installment plan credit. 0: No, 1: Yes}
\item{RENT}{Applicant rents. 0: No, 1: Yes}
\item{OWN_RES}{Applicant owns residence. 0: No, 1: Yes}
\item{NUM_CREDITS}{Number of existing credits at this bank}
\item{JOB}{Nature of job.
0: unemployed/ unskilled  - non-resident;
1: unskilled - resident;
2: skilled employee / official;
3: management/ self-employed/highly qualified employee/ officer}
\item{NUM_DEPENDENTS}{Number of people for whom liable to provide maintenance}
\item{TELEPHONE}{Applicant has phone in his or her name. 0: No, 1: Yes}
\item{FOREIGN}{Foreign worker. 0: No, 1: Yes}
\item{RESPONSE}{Credit rating is good. 0: No, 1: Yes}
}
}
\usage{
GermanCredit
}
\description{
Data
}
\examples{
head(GermanCredit)
}
\keyword{datasets}
