% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoteData.R
\name{getMovieLens}
\alias{getMovieLens}
\title{getMovieLense}
\source{
{The Movielens dataset is downloaded from \url{https://grouplens.org/datasets/movielens/}.
Dataset creation date: September 26, 2018). See \url{https://files.grouplens.org/datasets/movielens/ml-latest-small-README.html}
for the usage license.

F. Maxwell Harper and Joseph A. Konstan. 2015. The MovieLens Datasets: History and Context.
ACM Transactions on Interactive Intelligent Systems (TiiS) 5, 4: 19:1–19:19.
\url{https://doi.org/10.1145/2827872}}
}
\usage{
getMovieLens()
}
\value{
list containing the ratings and movies data

\describe{
\item{ratings: userId}{Anonymized MovieLens user ID. Each user has rated at least 20 movies.}
\item{ratings: movieId}{Movie ID consistent with that used on the MovieLens website (e.g., information for movie id 1 is available at \url{https://movielens.org/movies/1}).}
\item{ratings: rating}{Movie rating provided by user on a 5-star scale, with half-star increments (0.5 star - 5 stars)}
\item{ratings: timestamp}{Timestamp when rating was recorded. Time is in seconds since midnight Coordinated Universal Time (UTC) of January 1, 1970}
\item{movies: movieId}{Same as ratings data}
\item{movies: title}{Movie titles, including the year of release in parentheses}
\item{movies: genres}{Pipe-separated list of genres for corresponding movie title}
}
}
\description{
{Download Movielens 10k dataset}
}
\examples{
\dontrun{data <- mlba::getMovieLens()}
}
