% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoteData.R
\name{getMovieLense}
\alias{getMovieLense}
\title{getMovieLense}
\usage{
getMovieLense()
}
\value{
list containing the ratings and movies data

\describe{
\item{ratings: userId}{Anonymized MovieLens user ID. Each user has rated at least 20 movies.}
\item{ratings: movieId}{Movie ID consistent with that used on the MovieLens website (e.g., information for movie id 1 is available at \url{https://movielens.org/movies/1}).}
\item{ratings: rating}{Movie rating provided by user on a 5-star scale, with half-star increments (0.5 star - 5 stars)}
\item{ratings: timestamp}{Timestamp when rating was recorded. Time is in seconds since midnight Coordinated Universal Time (UTC) of January 1, 1970}
\item{movies: movieId}{Same as ratings data}
\item{movies: title}{Movie titles, including the year of release in parentheses}
\item{movies: genres}{Pipe-separated list of genres for corresponding movie title}
}
}
\description{
{Download Movielense data}
}
\examples{
\dontrun{data <- mlba::getMovieLense()}
}
