% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HairCareProduct}
\alias{HairCareProduct}
\title{Hair Care Product}
\format{
A data frame with 10000 observations and 8 variables:
\describe{
\item{Purchase}{}
\item{Age}{}
\item{Hair.Color}{}
\item{U.S..Region}{}
\item{Validation}{}
\item{Promotion_ord}{}
\item{Gender_ord}{}
\item{Residence_ord}{}
}
}
\usage{
HairCareProduct
}
\description{
Data
}
\examples{
head(HairCareProduct)
}
\keyword{datasets}
