% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LaptopSalesJanuary2008}
\alias{LaptopSalesJanuary2008}
\title{Laptop Sales January 2008}
\format{
A data frame with 7956 observations and 17 variables:
\describe{
\item{Date}{}
\item{Configuration}{}
\item{Customer.Postcode}{}
\item{Store.Postcode}{}
\item{Retail.Price}{}
\item{Screen.Size..Inches.}{}
\item{Battery.Life..Hours.}{}
\item{RAM..GB.}{}
\item{Processor.Speeds..GHz.}{}
\item{Integrated.Wireless.}{}
\item{HD.Size..GB.}{}
\item{Bundled.Applications.}{}
\item{OS.X.Customer}{}
\item{OS.Y.Customer}{}
\item{OS.X.Store}{}
\item{OS.Y.Store}{}
\item{CustomerStoreDistance}{}
}
}
\usage{
LaptopSalesJanuary2008
}
\description{
Data
}
\examples{
head(LaptopSalesJanuary2008)
}
\keyword{datasets}
