% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MovieLensMovies}
\alias{MovieLensMovies}
\title{MovieLensMovies}
\format{
A data frame with 9,742 observations and 3 variables:
\describe{
\item{movieId}{Same as ratings data}
\item{title}{Movie titles, including the year of release in parentheses}
\item{genres}{Pipe-separated list of genres for corresponding movie title}
}
}
\source{
{The Movielens dataset is downloaded from \url{https://grouplens.org/datasets/movielens/}.
Dataset creation date: September 26, 2018). See \url{https://files.grouplens.org/datasets/movielens/ml-latest-small-README.html}
for the usage license.

F. Maxwell Harper and Joseph A. Konstan. 2015. The MovieLens Datasets: History and Context.
ACM Transactions on Interactive Intelligent Systems (TiiS) 5, 4: 19:1–19:19.
\url{https://doi.org/10.1145/2827872}}
}
\usage{
MovieLensMovies
}
\description{
{Movies from the Movielens 10k dataset}
}
\examples{
head(MovieLensMovies)
}
\keyword{datasets}
