% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NaturalGasSales}
\alias{NaturalGasSales}
\title{NaturalGasSales}
\format{
A data frame with 16 observations and 2 variables:
\describe{
\item{Quarter}{}
\item{Gas.Sales}{}
}
}
\usage{
NaturalGasSales
}
\description{
Data
}
\examples{
NaturalGasSales
}
\keyword{datasets}
