% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ownerExample}
\alias{ownerExample}
\title{Owner example}
\format{
A data frame with 24 observations and 2 variables:
\describe{
\item{Class}{}
\item{Probability}{}
}
}
\usage{
ownerExample
}
\description{
Data
}
\examples{
head(ownerExample)
}
\keyword{datasets}
