% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pharmaceuticals}
\alias{Pharmaceuticals}
\title{Pharmaceuticals}
\format{
A data frame with 21 observations and 14 variables:
\describe{
\item{Symbol}{}
\item{Name}{}
\item{Market_Cap}{}
\item{Beta}{}
\item{PE_Ratio}{}
\item{ROE}{}
\item{ROA}{}
\item{Asset_Turnover}{}
\item{Leverage}{}
\item{Rev_Growth}{}
\item{Net_Profit_Margin}{}
\item{Median_Recommendation}{}
\item{Location}{}
\item{Exchange}{}
}
}
\source{
{Compiled from various web sources. Copyright 2016 Galit Shmueli and Peter Bruce}
}
\usage{
Pharmaceuticals
}
\description{
Data
}
\examples{
head(Pharmaceuticals)
}
\keyword{datasets}
