% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{regressionSummary}
\alias{regressionSummary}
\title{regressionSummary}
\usage{
regressionSummary(predicted, actual)
}
\arguments{
\item{predicted}{predicted values}

\item{actual}{actual values}
}
\value{
list containing the regression metrics ME, RMSE and MAE
}
\description{
Convenience function to calculate a variety of regression
metrics
}
\examples{
lm.SR <- lm(sr ~ pop15 + pop75 + dpi + ddpi, data = LifeCycleSavings)
regressionSummary(predict(lm.SR, LifeCycleSavings), LifeCycleSavings$sr)
}
