% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RidingMowers}
\alias{RidingMowers}
\title{Riding Mowers}
\format{
A data frame with 24 observations and 3 variables:
\describe{
\item{Income}{}
\item{Lot_Size}{}
\item{Ownership}{}
}
}
\usage{
RidingMowers
}
\description{
Data
}
\examples{
head(RidingMowers)
}
\keyword{datasets}
