% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SCstudents}
\alias{SCstudents}
\title{statistics.com Students}
\format{
A data frame with 1696 observations and 2 variables:
\describe{
\item{latitude}{}
\item{longitude}{}
}
}
\usage{
SCstudents
}
\description{
Data
}
\examples{
head(SCstudents)
}
\keyword{datasets}
