% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sept11Travel}
\alias{Sept11Travel}
\title{Sept11Travel}
\format{
A data frame with 172 observations and 4 variables:
\describe{
\item{Month}{}
\item{Air.RPM..000s.}{}
\item{Rail.PM}{}
\item{VMT..billions.}{}
}
}
\usage{
Sept11Travel
}
\description{
Data
}
\examples{
Sept11Travel
}
\keyword{datasets}
