% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ShampooSales}
\alias{ShampooSales}
\title{ShampooSales}
\format{
A data frame with 36 observations and 2 variables:
\describe{
\item{Month}{}
\item{Shampoo.Sales}{}
}
}
\usage{
ShampooSales
}
\description{
Data
}
\examples{
ShampooSales
}
\keyword{datasets}
