% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SouvenirSales}
\alias{SouvenirSales}
\title{SouvenirSales}
\format{
A data frame with 84 observations and 2 variables:
\describe{
\item{Date}{}
\item{Sales}{}
}
}
\usage{
SouvenirSales
}
\description{
Data
}
\examples{
SouvenirSales
}
\keyword{datasets}
