% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SP500}
\alias{SP500}
\title{SP500}
\format{
A data frame with 100 observations and 2 variables:
\describe{
\item{Date}{}
\item{Close}{}
}
}
\usage{
SP500
}
\description{
Data
}
\examples{
head(SP500)
}
\keyword{datasets}
