% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spambase}
\alias{spambase}
\title{spambase}
\format{
A data frame with 4601 observations and 58 variables:
\describe{
\item{make}{}
\item{address}{}
\item{all}{}
\item{W_3d}{}
\item{our}{}
\item{over}{}
\item{remove}{}
\item{internet}{}
\item{order}{}
\item{mail}{}
\item{receive}{}
\item{will}{}
\item{people}{}
\item{report}{}
\item{addresses}{}
\item{free}{}
\item{business}{}
\item{email}{}
\item{you}{}
\item{credit}{}
\item{your}{}
\item{font}{}
\item{W_000}{}
\item{money}{}
\item{hp}{}
\item{hpl}{}
\item{george}{}
\item{W_650}{}
\item{lab}{}
\item{labs}{}
\item{telnet}{}
\item{W_857}{}
\item{data}{}
\item{W_415}{}
\item{W_85}{}
\item{technology}{}
\item{W_1999}{}
\item{parts}{}
\item{pm}{}
\item{direct}{}
\item{cs}{}
\item{meeting}{}
\item{original}{}
\item{project}{}
\item{re.}{}
\item{edu}{}
\item{table}{}
\item{conference}{}
\item{C.}{}
\item{C..1}{}
\item{C..2}{}
\item{C..3}{}
\item{C..4}{}
\item{C..5}{}
\item{CAP_avg}{}
\item{CAP_long}{}
\item{CAP_tot}{}
\item{Spam}{}
}
}
\usage{
spambase
}
\description{
Data
}
\examples{
head(spambase)
}
\keyword{datasets}
