% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SystemAdministrators}
\alias{SystemAdministrators}
\title{System Administrator}
\format{
A data frame with 75 observations and 3 variables:
\describe{
\item{Experience}{}
\item{Training}{}
\item{Completed.task}{}
}
}
\usage{
SystemAdministrators
}
\description{
Data
}
\examples{
head(SystemAdministrators)
}
\keyword{datasets}
