% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TaxiCancellationCase}
\alias{TaxiCancellationCase}
\title{Taxi Cancellation Case}
\format{
A data frame with 10000 observations and 19 variables:
\describe{
\item{row.}{}
\item{user_id}{}
\item{vehicle_model_id}{}
\item{package_id}{}
\item{travel_type_id}{}
\item{from_area_id}{}
\item{to_area_id}{}
\item{from_city_id}{}
\item{to_city_id}{}
\item{from_date}{}
\item{to_date}{}
\item{online_booking}{}
\item{mobile_site_booking}{}
\item{booking_created}{}
\item{from_lat}{}
\item{from_long}{}
\item{to_lat}{}
\item{to_long}{}
\item{Car_Cancellation}{}
}
}
\source{
{Copyright 2016 statistics.com}
}
\usage{
TaxiCancellationCase
}
\description{
{The data are a randomly selected subset of the original
data, with 10,000 rows, one row for each booking of a taxi. There are
17 input variables, including user (customer) ID, vehicle model,
whether the booking was made online or via a mobile app, type of travel,
type of booking package, geographic information, and the date and
time of the scheduled trip. The target variable of interest is the
binary indicator of whether a ride was canceled.}
}
\examples{
head(TaxiCancellationCase)
}
\keyword{datasets}
