% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tayko}
\alias{Tayko}
\title{Tayko}
\format{
A data frame with 2000 observations and 25 variables:
\describe{
\item{sequence_number}{}
\item{US}{}
\item{source_a}{}
\item{source_c}{}
\item{source_b}{}
\item{source_d}{}
\item{source_e}{}
\item{source_m}{}
\item{source_o}{}
\item{source_h}{}
\item{source_r}{}
\item{source_s}{}
\item{source_t}{}
\item{source_u}{}
\item{source_p}{}
\item{source_x}{}
\item{source_w}{}
\item{Freq}{}
\item{last_update_days_ago}{}
\item{X1st_update_days_ago}{}
\item{Web.order}{}
\item{Gender.male}{}
\item{Address_is_res}{}
\item{Purchase}{}
\item{Spending}{}
}
}
\usage{
Tayko
}
\description{
Data
}
\examples{
head(Tayko)
}
\keyword{datasets}
