% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TinyData}
\alias{TinyData}
\title{TinyData}
\format{
A data frame with 6 observations and 4 variables:
\describe{
\item{Obs.}{}
\item{Fat}{}
\item{Salt}{}
\item{Acceptance}{}
}
}
\usage{
TinyData
}
\description{
Data
}
\examples{
head(TinyData)
}
\keyword{datasets}
