% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ToysRUsRevenues}
\alias{ToysRUsRevenues}
\title{ToysRUsRevenues}
\format{
A data frame with 16 observations and 4 variables:
\describe{
\item{Index}{}
\item{QuarterYear}{}
\item{Revenue.in.million...}{}
\item{Quarter}{}
}
}
\usage{
ToysRUsRevenues
}
\description{
Data
}
\examples{
ToysRUsRevenues
}
\keyword{datasets}
