% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UniversalBank}
\alias{UniversalBank}
\title{UniversalBank}
\format{
A data frame with 500 observations and 14 variables:
\describe{
\item{ID}{}
\item{Age}{}
\item{Experience}{}
\item{Income}{}
\item{ZIP.Code}{}
\item{Family}{}
\item{CCAvg}{}
\item{Education}{}
\item{Mortgage}{}
\item{Personal.Loan}{}
\item{Securities.Account}{}
\item{CD.Account}{}
\item{Online}{}
\item{CreditCard}{}
}
}
\usage{
UniversalBank
}
\description{
Data
}
\examples{
UniversalBank
}
\keyword{datasets}
