% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UniversalBankCase}
\alias{UniversalBankCase}
\title{UniversalBankCase}
\format{
A data frame with 5000 observations and 17 variables:
\describe{
\item{ID}{Customer ID}
\item{Age}{Customer's age in completed years}
\item{Experience}{#years of professional experience}
\item{Income}{Annual income of the customer ($000)}
\item{ZIP.Code}{Home Address ZIP code}
\item{Family}{Family size of the customer}
\item{CCAvg}{Avg. spending on credit cards per month ($000)}
\item{Education}{Education Level. 1: Undergrad; 2: Graduate; 3: Advanced/Professional}
\item{Mortgage}{Value of house mortgage if any. ($000)}
\item{Personal.Loan}{Did this customer accept the personal loan offered in the last campaign?}
\item{Securities.Account}{Does the customer have a securities account with the bank?}
\item{CD.Account}{Does the customer have a certificate of deposit (CD) account with the bank?}
\item{Online}{Does the customer use internet banking facilities?}
\item{CreditCard}{Does the customer use a credit card issued by UniversalBank?}
\item{Race}{0 = White/other (non-hisp), 1 = Black, 2 = Hispanic}
\item{Gender}{0 = Male, 1 = Female}
\item{Neighborhood}{1 = Urban, 0 = Other}
}
}
\source{
{Synthetic dataset courtesy of \url{statistics.com}.
Copyright Cytel, Inc. 2005, modifications Copyright Peter Bruce, 2022}
}
\usage{
UniversalBankCase
}
\description{
{A bank of small to moderate size is thinking of automating its
loan approval process to compete with internet-based financial service
companies.  A consultant has been engaged to develop a predictive model,
based on historical human-reviewed loans and their approval rates.  The
model would be used to approve or reject loans automatically.}
}
\examples{
UniversalBankCase
}
\keyword{datasets}
