% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Universities}
\alias{Universities}
\title{Universities}
\format{
A data frame with 1302 observations and 20 variables:
\describe{
\item{College.Name}{}
\item{State}{}
\item{Public..1...Private..2.}{}
\item{X..appli..rec.d}{}
\item{X..appl..accepted}{}
\item{X..new.stud..enrolled}{}
\item{X..new.stud..from.top.10.}{}
\item{X..new.stud..from.top.25.}{}
\item{X..FT.undergrad}{}
\item{X..PT.undergrad}{}
\item{in.state.tuition}{}
\item{out.of.state.tuition}{}
\item{room}{}
\item{board}{}
\item{add..fees}{}
\item{estim..book.costs}{}
\item{estim..personal..}{}
\item{X..fac..w.PHD}{}
\item{stud..fac..ratio}{}
\item{Graduation.rate}{}
}
}
\usage{
Universities
}
\description{
Data
}
\examples{
head(Universities)
}
\keyword{datasets}
