% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Utilities}
\alias{Utilities}
\title{Utilities}
\format{
A data frame with 22 observations and 9 variables:
\describe{
\item{Company}{}
\item{Fixed_charge}{}
\item{RoR}{}
\item{Cost}{}
\item{Load_factor}{}
\item{Demand_growth}{}
\item{Sales}{}
\item{Nuclear}{}
\item{Fuel_Cost}{}
}
}
\usage{
Utilities
}
\description{
Data
}
\examples{
Utilities
}
\keyword{datasets}
