% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Veerhoven}
\alias{Veerhoven}
\title{Veerhoven.}
\format{
A data frame with 159 observations and 5 variables:
\describe{
\item{Serial}{}
\item{Code}{}
\item{Nation}{}
\item{Score}{}
\item{X..surveys}{}
}
}
\usage{
Veerhoven
}
\description{
{Data measuring happiness of countries. according to a 2006 Gallup survey.}
}
\examples{
Veerhoven
}
\keyword{datasets}
