% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WalMartStock}
\alias{WalMartStock}
\title{WalMartStock}
\format{
A data frame with 248 observations and 2 variables:
\describe{
\item{Date}{}
\item{Close}{}
}
}
\usage{
WalMartStock
}
\description{
Data
}
\examples{
WalMartStock
}
\keyword{datasets}
