% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WestRoxbury}
\alias{WestRoxbury}
\title{West Roxbury.}
\format{
A data frame with 5802 observations and 14 variables:
\describe{
\item{TOTAL.VALUE}{}
\item{TAX}{}
\item{LOT.SQFT}{}
\item{YR.BUILT}{}
\item{GROSS.AREA}{}
\item{LIVING.AREA}{}
\item{FLOORS}{}
\item{ROOMS}{}
\item{BEDROOMS}{}
\item{FULL.BATH}{}
\item{HALF.BATH}{}
\item{KITCHEN}{}
\item{FIREPLACE}{}
\item{REMODEL}{}
}
}
\usage{
WestRoxbury
}
\description{
Data
}
\examples{
WestRoxbury
}
\keyword{datasets}
