% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Wine}
\alias{Wine}
\title{Wine.}
\format{
A data frame with 178 observations and 14 variables:
\describe{
\item{Type}{}
\item{Alcohol}{}
\item{Malic_Acid}{}
\item{Ash}{}
\item{Ash_Alcalinity}{}
\item{Magnesium}{}
\item{Total_Phenols}{}
\item{Flavanoids}{}
\item{Nonflavanoid_Phenols}{}
\item{Proanthocyanins}{}
\item{Color_Intensity}{}
\item{Hue}{}
\item{OD280_OD315}{}
\item{Proline}{}
}
}
\usage{
Wine
}
\description{
Data
}
\examples{
Wine
}
\keyword{datasets}
