% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accidentsnn}
\alias{accidentsnn}
\title{accidentsnn}
\format{
A data frame with 999 observations and 5 variables:
\describe{
\item{ALCHL_I}{}
\item{PROFIL_I_R}{}
\item{SUR_COND}{}
\item{VEH_INVL}{}
\item{MAX_SEV_IR}{}
}
}
\usage{
accidentsnn
}
\description{
Data
}
\examples{
head(accidentsnn)
}
\keyword{datasets}
