% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Airfares}
\alias{Airfares}
\title{Airfares}
\format{
A data frame with 638 observations and 18 variables:
\describe{
\item{S_CODE}{starting airport's code}
\item{S_CITY}{starting city}
\item{E_CODE}{ending airport's code}
\item{E_CITY}{ending city}
\item{COUPON}{average number of coupons (a one-coupon flight is a non-stop flight, a two-coupon flight is a one stop flight, etc.) for that route}
\item{NEW}{number of new carriers entering that route between Q3-96 and Q2-97}
\item{VACATION}{whether a vacation route (Yes) or not (No); Florida and Las Vegas routes are generally considered vacation routes}
\item{SW}{whether Southwest Airlines serves that route (Yes) or not (No)}
\item{HI}{Herfindel Index – measure of market concentration (refer to BMGT 681)}
\item{S_INCOME}{starting city's average personal income}
\item{E_INCOME}{ending city's average personal income}
\item{S_POP}{starting city's population}
\item{E_POP}{ending city's population}
\item{SLOT}{whether either endpoint airport is slot controlled or not; this is a measure of airport congestion}
\item{GATE}{whether either endpoint airport has gate constraints or not; this is another measure of airport congestion}
\item{DISTANCE}{distance between two endpoint airports in miles}
\item{PAX}{number of passengers on that route during period of data collection}
\item{FARE}{average fare on that route}
}
}
\source{
{Copyright 2016 Galit Shmueli and Peter Bruce}
}
\usage{
Airfares
}
\description{
Airfares
}
\examples{
head(Airfares)
}
\keyword{datasets}
