% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ApplianceShipments}
\alias{ApplianceShipments}
\title{Applicance Shipments}
\format{
A data frame with 20 observations and 2 variables:
\describe{
\item{Quarter}{}
\item{Shipments}{}
}
}
\source{
{Data courtesy Ken Black}
}
\usage{
ApplianceShipments
}
\description{
The series of quarterly shipments (in millions of dollars) of US household
appliances between 1985 and 1989.
}
\examples{
head(ApplianceShipments)
}
\keyword{datasets}
