% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Banks}
\alias{Banks}
\title{Banks}
\format{
A data frame with 20 observations and 5 variables:
\describe{
\item{Obs}{}
\item{Financial.Condition}{1 = financially weak; 0 = financially strong}
\item{TotCap.Assets}{}
\item{TotExp.Assets}{}
\item{TotLns.Lses.Assets}{}
}
}
\usage{
Banks
}
\description{
Data
}
\examples{
Banks
}
\keyword{datasets}
