% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CanadianWorkHours}
\alias{CanadianWorkHours}
\title{CanadianWorkHours}
\format{
A data frame with 35 observations and 2 variables:
\describe{
\item{Year}{}
\item{Hours}{average annual number of weekly hours spent by Canadian manufacturing workers}
}
}
\source{
{Ken Black (used by permission)}
}
\usage{
CanadianWorkHours
}
\description{
Data
}
\examples{
CanadianWorkHours
}
\keyword{datasets}
