% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CharlesBookClub}
\alias{CharlesBookClub}
\title{Charles Book Club}
\format{
A data frame with 4000 observations and 24 variables:
\describe{
\item{Seq.}{Sequence number in the sample}
\item{ID.}{ID# in the full dataset}
\item{Gender}{0=male, 1=female}
\item{M}{Monetary - total money spent on books}
\item{R}{Recency - Months since last purchase}
\item{F}{Frequency - Total number of purchases}
\item{FirstPurch}{Months since first purchase}
\item{ChildBks}{this and following - book categories}
\item{YouthBks}{}
\item{CookBks}{}
\item{DoItYBks}{}
\item{RefBks}{}
\item{ArtBks}{}
\item{GeogBks}{}
\item{ItalCook}{}
\item{ItalAtlas}{}
\item{ItalArt}{}
\item{Florence}{}
\item{Related.Purchase}{Number of related books purchased}
\item{Mcode}{Recoding of M - see case description in DMBA and MLBA}
\item{Rcode}{Recoding of R - see case description in DMBA and MLBA}
\item{Fcode}{Recoding of F - see case description in DMBA and MLBA}
\item{Yes_Florence}{}
\item{No_Florence}{}
}
}
\source{
{Adapted with permission from The Bookbinders Club, prepared by Nissan Levin and Jacob Zahavi.}
}
\usage{
CharlesBookClub
}
\description{
Data
}
\examples{
head(CharlesBookClub)
}
\keyword{datasets}
