% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CourseTopics}
\alias{CourseTopics}
\title{CourseTopics}
\format{
A data frame with 365 observations and 8 variables:
\describe{
\item{Intro}{}
\item{DataMining}{}
\item{Survey}{}
\item{Cat.Data}{}
\item{Regression}{}
\item{Forecast}{}
\item{DOE}{}
\item{SW}{}
}
}
\source{
{Copyright 2016 Galit Shmueli and Peter Bruce}
}
\usage{
CourseTopics
}
\description{
Course topics at statistics.com (each row is a customer, column heads are
topics taken [1] or not taken [0] by that customer)
}
\examples{
head(CourseTopics)
}
\keyword{datasets}
