% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Faceplate}
\alias{Faceplate}
\title{Faceplate}
\format{
A data frame with 10 observations and 7 variables:
\describe{
\item{Transaction}{}
\item{Red}{}
\item{White}{}
\item{Blue}{}
\item{Orange}{}
\item{Green}{}
\item{Yellow}{}
}
}
\source{
{Copyright 2016 Galit Shmueli and Peter Bruce}
}
\usage{
Faceplate
}
\description{
Synthetic Data on Purchases of Phone Faceplates.
}
\examples{
Faceplate
}
\keyword{datasets}
