% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FarmAds}
\alias{FarmAds}
\title{FarmAds}
\format{
A data frame with 4143 observations and 2 variables:
\describe{
\item{label}{1: ad is relevant; -1 ad is not relevant}
\item{text}{text of ad pre-processed into a list of words. The location
of the word in the ad is in some cases indicated by a prefix.}
}
}
\usage{
FarmAds
}
\description{
Data on advertisements posted at a website that caters to the needs of
a specific farming community. Each ad is in a row, and each ad labeled
as either -1 (not relevant) or 1 (relevant). The goal is to develop a
predictive model that can classify ads automatically.
}
\details{
Data
}
\examples{
FarmAds
}
\keyword{datasets}
