% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GDP}
\alias{GDP}
\title{GDP}
\format{
A data frame with 264 observations and 5 variables:
\describe{
\item{Country.Name}{}
\item{Country.Code}{}
\item{Indicator.Name}{}
\item{Indicator.Code}{}
\item{GDP2015}{Gross domestic product of the countries.}
}
}
\source{
{Data from Veenhoven's world database of happiness.
\url{http://data.worldbank.org/indicator/NY.GDP.MKTP.CD}.
World Development Indicators.  }
}
\usage{
GDP
}
\description{
Data
}
\examples{
head(GDP)
}
\keyword{datasets}
