% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LaptopSales}
\alias{LaptopSales}
\title{Laptop Sales}
\format{
A data frame with 297,572 observations and 16 variables:
\describe{
\item{Date}{purchase date}
\item{Configuration}{A numerical code representing a combination of screen
size, battery life, RAM, etc. Each code corresponds to a particular combination.}
\item{Customer.Postcode}{postcode in London of the customer}
\item{Store.Postcode}{postcode in London of the store}
\item{Retail.Price}{price of laptop in GBP}
\item{Screen.Size..Inches.}{screen size of laptop (Inches)}
\item{Battery.Life..Hours.}{battery life of laptop (Hours)}
\item{RAM..GB.}{RAM size of laptop(GB)}
\item{Processor.Speeds..GHz.}{processor speed of laptop (GHz)}
\item{Integrated.Wireless.}{whether the laptop has integrated wireless or not}
\item{HD.Size..GB.}{HD size of laptop (GB)}
\item{Bundled.Applications.}{whether the laptop comes with bundled applications or not}
\item{customer.X}{X geo coordinates for customer location.}
\item{customer.Y}{Y geo coordinates for customer location.}
\item{store.X}{X geo coordinates for store location}
\item{store.Y}{Y geo coordinates for store location}
}
}
\source{
{The laptop sales data were part of the ENBIS 2009 Challenge
in Industrial Statistics}
}
\usage{
LaptopSales
}
\description{
{The laptop sales data were part of the ENBIS 2009 Challenge in
Industrial Statistics.

This is the full Laptop sales dataset. It includes only the Jan 2008
sales (the complete dataset includes the entire 2008 sales).}

Data
}
\examples{
head(LaptopSales)
}
\keyword{datasets}
