% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liftExample}
\alias{liftExample}
\title{Lift example}
\format{
A data frame with 24 observations and 2 variables:
\describe{
\item{prob}{}
\item{actual}{}
}
}
\usage{
liftExample
}
\description{
{Synthetic dataset to demonstrate lift}
}
\examples{
head(liftExample)
}
\keyword{datasets}
