% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LiftExample}
\alias{LiftExample}
\title{LiftExample}
\format{
A data frame with 24 observations and 2 variables:
\describe{
\item{prob}{}
\item{actual}{}
}
}
\usage{
LiftExample
}
\description{
{Synthetic dataset to demonstrate lift}
}
\examples{
head(LiftExample)
}
\keyword{datasets}
