% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NYPDMotorVehicleCollisions}
\alias{NYPDMotorVehicleCollisions}
\title{MovieLenseMovies}
\format{
A data frame with 9,742 observations and 3 variables:
\describe{
\item{movieId}{Same as ratings data}
\item{title}{Movie titles, including the year of release in parentheses}
\item{genres}{Pipe-separated list of genres for corresponding movie title}
}

A data frame with 100,836 observations and 4 variables:
\describe{
\item{userId}{Anonymized MovieLens user ID. Each user has rated at least 20 movies.}
\item{movieId}{Movie ID consistent with that used on the MovieLens website (e.g., information for movie id 1 is available at \url{https://movielens.org/movies/1}).}
\item{rating}{Movie rating provided by user on a 5-star scale, with half-star increments (0.5 star - 5 stars)}
\item{timestamp}{Timestamp when rating was recorded. Time is in seconds since midnight Coordinated Universal Time (UTC) of January 1, 1970}
}

A data frame with 1000 observations and 29 variables:
\describe{
\item{DATE}{Occurrence date of collision}
\item{TIME}{Occurrence time of collision}
\item{BOROUGH}{Borough where collision occurred}
\item{ZIP.CODE}{Postal code of incident occurrence}
\item{LATITUDE}{Latitude coordinate for Global Coordinate System, WGS 1984, decimal degrees (EPSG 4326)}
\item{LONGITUDE}{Longitude coordinate for Global Coordinate System, WGS 1984, decimal degrees (EPSG 4326)}
\item{LOCATION}{Latitude , Longitude pair}
\item{ON.STREET.NAME}{Street on which the collision occurred}
\item{CROSS.STREET.NAME}{Nearest cross street to the collision}
\item{OFF.STREET.NAME}{Street address if known}
\item{NUMBER.OF.PERSONS.INJURED}{Number of persons injured}
\item{NUMBER.OF.PERSONS.KILLED}{Number of persons killed}
\item{NUMBER.OF.PEDESTRIANS.INJURED}{Number of pedestrians injured}
\item{NUMBER.OF.PEDESTRIANS.KILLED}{Number of pedestrians killed}
\item{NUMBER.OF.CYCLIST.INJURED}{Number of cyclists injured}
\item{NUMBER.OF.CYCLIST.KILLED}{Number of cyclists killed}
\item{NUMBER.OF.MOTORIST.INJURED}{Number of vehicle occupants injured}
\item{NUMBER.OF.MOTORIST.KILLED}{Number of vehicle occupants killed}
\item{CONTRIBUTING.FACTOR.VEHICLE.1}{Factors contributing to the collision for designated vehicle}
\item{CONTRIBUTING.FACTOR.VEHICLE.2}{Factors contributing to the collision for designated vehicle}
\item{CONTRIBUTING.FACTOR.VEHICLE.3}{Factors contributing to the collision for designated vehicle}
\item{CONTRIBUTING.FACTOR.VEHICLE.4}{Factors contributing to the collision for designated vehicle}
\item{CONTRIBUTING.FACTOR.VEHICLE.5}{Factors contributing to the collision for designated vehicle}
\item{UNIQUE.KEY}{Unique record code generated by system. Primary Key for Crash table.}
\item{VEHICLE.TYPE.CODE.1}{Type of vehicle based on the selected vehicle category (ATV, bicycle, car/suv, ebike, escooter, truck/bus, motorcycle, other)}
\item{VEHICLE.TYPE.CODE.2}{Type of vehicle based on the selected vehicle category (ATV, bicycle, car/suv, ebike, escooter, truck/bus, motorcycle, other)}
\item{VEHICLE.TYPE.CODE.3}{Type of vehicle based on the selected vehicle category (ATV, bicycle, car/suv, ebike, escooter, truck/bus, motorcycle, other)}
\item{VEHICLE.TYPE.CODE.4}{Type of vehicle based on the selected vehicle category (ATV, bicycle, car/suv, ebike, escooter, truck/bus, motorcycle, other)}
\item{VEHICLE.TYPE.CODE.5}{Type of vehicle based on the selected vehicle category (ATV, bicycle, car/suv, ebike, escooter, truck/bus, motorcycle, other)}
}
}
\source{
{The Movielens dataset is downloaded from \url{https://grouplens.org/datasets/movielens/}.
Dataset creation date: September 26, 2018). See \url{https://files.grouplens.org/datasets/movielens/ml-latest-small-README.html}
for the usage license.

F. Maxwell Harper and Joseph A. Konstan. 2015. The MovieLens Datasets: History and Context.
ACM Transactions on Interactive Intelligent Systems (TiiS) 5, 4: 19:1–19:19.
\url{https://doi.org/10.1145/2827872}}

{The Movielens dataset is downloaded from \url{https://grouplens.org/datasets/movielens/}.
Dataset creation date: September 26, 2018). See \url{https://files.grouplens.org/datasets/movielens/ml-latest-small-README.html}
for the usage license.

F. Maxwell Harper and Joseph A. Konstan. 2015. The MovieLens Datasets: History and Context.
ACM Transactions on Interactive Intelligent Systems (TiiS) 5, 4: 19:1–19:19.
\url{https://doi.org/10.1145/2827872}}

{Random sample of the NYPD Motor Vehicle Collisions Dataset.
(\url{https://data.cityofnewyork.us/Public-Safety/NYPD-Motor-Vehicle-Collisions/h9gi-nx95})}
}
\usage{
NYPDMotorVehicleCollisions
}
\description{
{Movies from the Movielens 10k dataset}

{Ratings from the Movielens 10k dataset}

{The Motor Vehicle Collisions crash table contains details on
the crash event. Each row represents a crash event. The Motor Vehicle
Collisions data tables contain information from all police reported motor
vehicle collisions in NYC. This dataset is a random subset of 1000 entries
of the NYPD motor vehicle collisions dataset}
}
\examples{
\dontrun{head(MovieLenseMovies)}
MovieLenseRatings

\dontrun{head(MovieLenseRatings)}
NYPD Motor Vehicle Collisions

head(NYPDMotorVehicleCollisions)
}
\keyword{datasets}
