% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OwnerExample}
\alias{OwnerExample}
\title{OwnerExample}
\format{
A data frame with 24 observations and 2 variables:
\describe{
\item{Class}{}
\item{Probability}{}
}
}
\usage{
OwnerExample
}
\description{
Data
}
\examples{
head(OwnerExample)
}
\keyword{datasets}
