% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RidingMowers}
\alias{RidingMowers}
\title{Riding Mowers}
\format{
A data frame with 24 observations and 3 variables:
\describe{
\item{Income}{Annual income in $000}
\item{Lot_Size}{In thousands of sq. feet}
\item{Ownership}{Whether the resident owns a riding mower or not}
}
}
\source{
{Data courtesy of Dean Wichern}
}
\usage{
RidingMowers
}
\description{
Data
}
\examples{
head(RidingMowers)
}
\keyword{datasets}
