% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sept11Travel}
\alias{Sept11Travel}
\title{Sept11Travel}
\format{
A data frame with 172 observations and 4 variables:
\describe{
\item{Month}{}
\item{Air.RPM..000s.}{Air revenue passenger miles (1 RMP is one revenue passenger carried for one mile)}
\item{Rail.PM}{Rail passenger miles}
\item{VMT..billions.}{Vehicle miles traveled}
}
}
\source{
{Bureau of Transportation Statistics:
  \url{https://www.bts.gov/archive/publications/estimated_impacts_of_9_11_on_us_travel/index}}
}
\usage{
Sept11Travel
}
\description{
{Estimated Impacts of September 11th on US Travel}
}
\examples{
Sept11Travel
}
\keyword{datasets}
