% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ShampooSales}
\alias{ShampooSales}
\title{ShampooSales}
\format{
A data frame with 36 observations and 2 variables:
\describe{
\item{Month}{}
\item{Shampoo.Sales}{}
}
}
\source{
{Time Series Data Library, \url{https://pkg.yangzhuoranyang.com/tsdl/}}
}
\usage{
ShampooSales
}
\description{
{Data on the monthly sales of a certain shampoo over a 3-year period.}
}
\examples{
ShampooSales
}
\keyword{datasets}
