% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SouvenirSales}
\alias{SouvenirSales}
\title{SouvenirSales}
\format{
A data frame with 84 observations and 2 variables:
\describe{
\item{Date}{}
\item{Sales}{}
}
}
\source{
{Time Series Data Library, \url{https://pkg.yangzhuoranyang.com/tsdl/}}
}
\usage{
SouvenirSales
}
\description{
{Monthly sales for a souvenir shop at a beach resort town in Queensland, Australia, between 1995–2001.}
}
\examples{
SouvenirSales
}
\keyword{datasets}
