% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Spambase}
\alias{Spambase}
\title{Spambase}
\format{
A data frame with 4601 observations and 58 variables:
\describe{
\item{make}{}
\item{address}{}
\item{all}{}
\item{W_3d}{}
\item{our}{}
\item{over}{}
\item{remove}{}
\item{internet}{}
\item{order}{}
\item{mail}{}
\item{receive}{}
\item{will}{}
\item{people}{}
\item{report}{}
\item{addresses}{}
\item{free}{}
\item{business}{}
\item{email}{}
\item{you}{}
\item{credit}{}
\item{your}{}
\item{font}{}
\item{W_000}{}
\item{money}{}
\item{hp}{}
\item{hpl}{}
\item{george}{}
\item{W_650}{}
\item{lab}{}
\item{labs}{}
\item{telnet}{}
\item{W_857}{}
\item{data}{}
\item{W_415}{}
\item{W_85}{}
\item{technology}{}
\item{W_1999}{}
\item{parts}{}
\item{pm}{}
\item{direct}{}
\item{cs}{}
\item{meeting}{}
\item{original}{}
\item{project}{}
\item{re.}{}
\item{edu}{}
\item{table}{}
\item{conference}{}
\item{C.}{C;}
\item{C..1}{C(}
\item{C..2}{C[}
\item{C..3}{C!}
\item{C..4}{C$}
\item{C..5}{C#}
\item{CAP_avg}{average length of uninterrupted sequences of capital letters}
\item{CAP_long}{length of longest uninterrupted sequence of capital letters}
\item{CAP_tot}{total number of capital letters in the e-mail}
\item{Spam}{1 = spam, 0 = not spam}
}
}
\usage{
Spambase
}
\description{
{ Each of the words below are columns in the data and the values represent
\% of words in the e-mail that match that particular word. For example,
make represent \% of words in the e-mail that match "make".}
}
\examples{
head(Spambase)
}
\keyword{datasets}
