% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SystemAdministrators}
\alias{SystemAdministrators}
\title{System Administrator}
\format{
A data frame with 75 observations and 3 variables:
\describe{
\item{Experience}{measures months of full-time system administrator experience}
\item{Training}{measures the number of relevant training credits}
\item{Completed.task}{either Yes or No, according to whether or not the
administrator completed the tasks}
}
}
\source{
{Samprit Chatterjee}
}
\usage{
SystemAdministrators
}
\description{
Data
}
\examples{
head(SystemAdministrators)
}
\keyword{datasets}
