% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TaxiCancellationCase}
\alias{TaxiCancellationCase}
\title{Taxi Cancellation Case}
\format{
A data frame with 10000 observations and 19 variables:
\describe{
\item{row.}{}
\item{user_id}{the ID of the customer (based on mobile number)}
\item{vehicle_model_id}{vehicle model type.}
\item{package_id}{type of package (1=4hrs & 40kms, 2=8hrs & 80kms, 3=6hrs & 60kms, 4= 10hrs & 100kms, 5=5hrs & 50kms, 6=3hrs & 30kms, 7=12hrs & 120kms)}
\item{travel_type_id}{type of travel (1=long distance, 2= point to point, 3= hourly rental).}
\item{from_area_id}{unique identifier of area. Applicable only for point-to-point travel and packages}
\item{to_area_id}{unique identifier of area. Applicable only for point-to-point travel}
\item{from_city_id}{unique identifier of city}
\item{to_city_id}{unique identifier of city (only for intercity)}
\item{from_date}{time stamp of requested trip start}
\item{to_date}{time stamp of trip end}
\item{online_booking}{if booking was done on desktop website}
\item{mobile_site_booking}{if booking was done on mobile website}
\item{booking_created}{time stamp of booking}
\item{from_lat}{latitude of from area}
\item{from_long}{longitude of from area}
\item{to_lat}{latitude of to area}
\item{to_long}{longitude of to area}
\item{Car_Cancellation}{1=trip cancelled; 0=trip not cancelled}
}
}
\source{
{Copyright 2016 statistics.com}
}
\usage{
TaxiCancellationCase
}
\description{
{The data are a randomly selected subset of the original
data, with 10,000 rows, one row for each booking of a taxi. There are
17 input variables, including user (customer) ID, vehicle model,
whether the booking was made online or via a mobile app, type of travel,
type of booking package, geographic information, and the date and
time of the scheduled trip. The target variable of interest is the
binary indicator of whether a ride was canceled.}
}
\examples{
head(TaxiCancellationCase)
}
\keyword{datasets}
