% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Universities}
\alias{Universities}
\title{Universities}
\format{
A data frame with 1302 observations and 20 variables:
\describe{
\item{College.Name}{}
\item{State}{}
\item{Public..1...Private..2.}{}
\item{X..appli..rec.d}{}
\item{X..appl..accepted}{}
\item{X..new.stud..enrolled}{}
\item{X..new.stud..from.top.10.}{}
\item{X..new.stud..from.top.25.}{}
\item{X..FT.undergrad}{}
\item{X..PT.undergrad}{}
\item{in.state.tuition}{}
\item{out.of.state.tuition}{}
\item{room}{}
\item{board}{}
\item{add..fees}{}
\item{estim..book.costs}{}
\item{estim..personal..}{}
\item{X..fac..w.PHD}{}
\item{stud..fac..ratio}{}
\item{Graduation.rate}{}
}
}
\source{
{Copyright 2016 Galit Shmueli and Peter Bruce. Compiled from US News and
World Report rankings on 1302American Colleges and Universities}
}
\usage{
Universities
}
\description{
{The dataset on American college and university rankings contains
information on 1302 American colleges and universities oﬀering an undergraduate
program. For each university, there are 17 measurements that include
continuous measurements (such as tuition and graduation rate) and categorical
measurements (such as location by state and whether it is a private or a public school).}
}
\examples{
head(Universities)
}
\keyword{datasets}
